# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: PrestaRapi - Sistema de Gestión de Préstamos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /funders:
    get:
      operationId: getFunders
      tags: [funders]
      summary: List funders
      parameters:
        - name: agencyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of funders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Funder"
    post:
      operationId: createFunder
      tags: [funders]
      summary: Create funder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunderRequest"
      responses:
        "201":
          description: Funder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funder"
  /funders/{id}:
    put:
      operationId: updateFunder
      tags: [funders]
      summary: Update funder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunderRequest"
      responses:
        "200":
          description: Funder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funder"
    delete:
      operationId: deleteFunder
      tags: [funders]
      summary: Soft-delete funder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /funders/{id}/transaction:
    post:
      operationId: funderTransaction
      tags: [funders]
      summary: Add capital or withdrawal for funder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FunderTransactionRequest"
      responses:
        "200":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funder"
  /holidays:
    get:
      operationId: getHolidays
      tags: [holidays]
      summary: List holidays
      responses:
        "200":
          description: List of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create holiday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHolidayRequest"
      responses:
        "201":
          description: Holiday created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    put:
      operationId: updateHoliday
      tags: [holidays]
      summary: Update holiday
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHolidayRequest"
      responses:
        "200":
          description: Holiday updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete holiday
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Funder:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        capital:
          type: number
        rateToPayPercent:
          type: number
        agencyId:
          type: integer
        agencyName:
          type: string
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - capital
        - rateToPayPercent
        - agencyId
        - periodicity
        - active
        - createdAt
    CreateFunderRequest:
      type: object
      properties:
        name:
          type: string
        capital:
          type: number
        rateToPayPercent:
          type: number
        agencyId:
          type: integer
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
      required:
        - name
        - capital
        - rateToPayPercent
        - agencyId
        - periodicity
    FunderTransactionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [increase, withdrawal]
        amount:
          type: number
        note:
          type: string
        proof:
          type: string
      required:
        - type
        - amount
    Holiday:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amortization:
          type: string
          enum: [none, advance, delay]
      required:
        - id
        - date
        - description
        - amortization
    CreateHolidayRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amortization:
          type: string
          enum: [none, advance, delay]
      required:
        - date
        - description
        - amortization
...